+++ src/fr/helmdefense/model/entities/attackers/bosses/MordorTower.java
@@ -1,6 +1,6 @@
 package fr.helmdefense.model.entities.attackers.bosses;
 
-import fr.helmdefense.model.entities.utils.Location;
+import fr.helmdefense.model.entities.utils.coords.Location;
 
 public class MordorTower extends TwoTowers {
 
@@ -8,7 +8,7 @@ public class MordorTower extends TwoTowers {
 		super(location);
 	}
 	
-	public MordorTower(int x, int y) {
+	public MordorTower(double x, double y) {
 		this(new Location(x, y));
 	}
 }
