+++ src/fr/helmdefense/model/entities/defenders/Defender.java
@@ -1,7 +1,7 @@
 package fr.helmdefense.model.entities.defenders;
 
 import fr.helmdefense.model.entities.Entity;
-import fr.helmdefense.model.entities.utils.Location;
+import fr.helmdefense.model.entities.utils.coords.Location;
 
 public abstract class Defender extends Entity {
 	
@@ -9,7 +9,7 @@ public abstract class Defender extends Entity {
 		super(loc);
 	}
 	
-	public Defender(int x, int y) {
+	public Defender(double x, double y) {
 		this(new Location(x,y));
 	}
 }
