+++ src/fr/helmdefense/model/entities/defenders/HumanWarrior.java
@@ -1,17 +1,13 @@
 package fr.helmdefense.model.entities.defenders;
 
-import fr.helmdefense.model.entities.utils.Location;
+import fr.helmdefense.model.entities.utils.coords.Location;
 
 public class HumanWarrior extends Defender {
-	
 	public HumanWarrior(Location loc) {
 		super(loc);
 	}
 	
-	public HumanWarrior(int x, int y) {
+	public HumanWarrior(double x, double y) {
 		this(new Location(x,y));
 	}
-	
-	
-
 }
