+++ src/fr/helmdefense/model/entities/defenders/heros/Legolas.java
@@ -1,6 +1,6 @@
 package fr.helmdefense.model.entities.defenders.heros;
 
-import fr.helmdefense.model.entities.utils.Location;
+import fr.helmdefense.model.entities.utils.coords.Location;
 
 
 public class Legolas extends Hero {
@@ -9,7 +9,7 @@ public class Legolas extends Hero {
 		super(loc);
 	}
 	
-	public Legolas(int x, int y) {
+	public Legolas(double x, double y) {
 		this(new Location(x, y));
 	}
 }
